pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tajbench"
version = "0.1.0"
description = "TajBench — a rigorous open benchmark for evaluating frontier LLMs on rice population genomics reasoning"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
  { name = "True Cultivar" },
]
keywords = ["genomics", "LLM", "benchmark", "population-genetics", "rice", "plant-genomics"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Science/Research",
  "Topic :: Scientific/Engineering :: Bio-Informatics",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
]

dependencies = [
  "anthropic>=0.40.0",
  "openai>=1.50.0",
  "google-generativeai>=0.8.0",
  "requests>=2.32.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.6",
]

[project.scripts]
tajb = "harness.runner:main"

[tool.hatch.build.targets.wheel]
packages = ["benchmark", "harness"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
